docker/app/requirements.txt
flask==2.0.3
Werkzeug==2.0.3
python-docx
pandas
numpy


